moved sumatrix to dmatrix
